pyproject.toml
[project]
name = "langgraph-orches"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Saim-Hassan786", email = "saimaiengineer@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "ipython>=8.32.0",
    "langchain-google-genai>=2.0.9",
    "langgraph>=0.2.74",
    "python-dotenv>=1.0.1",
]

[project.scripts]
langgraph-orches = "langgraph_orches:main"
orch = "langgraph_orches.orches:kickoff"
claud = "langgraph_orches.orches2:kickoff"
eval = "langgraph_orches.eval:evalit"
react = "langgraph_orches.react_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
